backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
faiss-cpu
sentence-transformers
numpy